Removed Temperaturelist file. It was adding no functionallity and just acting a a container that decrease ease of styling
