+++ src/App.css
@@ -28,16 +28,16 @@ table {
 	background-color: #4caf50;
 	color: white;
 }
-
+/* 
 .graph-table-con {
 	display: grid;
-	/* grid-template-rows: 1fr 1fr;
-    grid-template-columns: 1000px; */
+	grid-template-rows: 1fr 1fr;
+    grid-template-columns: 1000px;
 	grid-template-columns: 1fr 1fr;
-	grid-template-rows: 400px;
+	grid-template-rows: 400px 20;
 	column-gap: 20px;
 	row-gap: 20px;
-}
+} */
 
 .graph {
 	grid-column: 1 / span 1;
@@ -58,7 +58,7 @@ tr:nth-child(even) {
 	background-color: #f2f2f2;
 }
 
-.quick-stats-con {
+/* .quick-stats-con {
 	display: flex;
 	flex-wrap: wrap;
 	justify-content: space-around;
@@ -66,7 +66,7 @@ tr:nth-child(even) {
 	grid-template-columns: 1fr 1fr 1fr 1fr;
 	grid-template-rows: 1fr 1fr 1fr 1fr;
 	grid-template-rows: 200px; */
-}
+/* } */
 /* Login Modal Attribution: https://www.geeksforgeeks.org/html-responsive-modal-login-form/ & General Assembly */
 #modal-login {
 	background-color: rgba(0, 0, 0, 0.4);
@@ -148,4 +148,10 @@ img.avatar {
 .card:hover {
   box-shadow: 0 8px 16px 0 rgba(0,0,0,0.2);
   width: 15.5%;
+}
+
+.table-graph-container{
+	display: grid;
+	grid-template-columns: 1fr 1fr;
+	
 }
