+++ src/components/aquarium_status/temperature/DateRangeForm.js
@@ -1,9 +1,34 @@
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
+import moment from 'moment';
+import axios from 'axios'
 
-const DateRangeForm = ({ date, setDate, refresh, setRefresh }) => {
+const m = moment();
+let current_formatted_date = `${m.date()}-${m.month() + 1}-${m.year()}`;
+
+const DateRangeForm = ({setTemperatureData}) => {
+	//Added
+	const [date, setDate] = useState({
+		start: current_formatted_date,
+		end: current_formatted_date,
+	});
+	const [refresh, setRefresh] = useState(false);
 	function handleDateChange(event) {
 		setDate({ ...date, [event.target.id]: event.target.value });
 	}
+
+	const url = 'https://mighty-lake-45709.herokuapp.com/temperatures/range/';
+	// const url = 'http://localhost:8000/temperatures/range/';
+	// const url = './temp_seeds.json';
+	useEffect(() => {
+		axios
+			.get(url, { params: { start: date.start, end: date.end } })
+			.then((res) => {
+				setTemperatureData(res.data);
+			})
+			.catch(console.error);
+	}, [refresh]);
+
+	//^^Added
 	function handleDateRangeSubmint(event) {
 		event.preventDefault();
 		refresh ? setRefresh(false) : setRefresh(true);
